fix: implement access log in middleware, allows adding timing info and solves cloudwatch support
