+++ vmaas/common/probes.py
@@ -7,8 +7,8 @@ doesn't return a Histogram, but rather a LabelWrapper.
 
 from prometheus_client import Histogram, Counter
 
-REQUEST_TIME = Histogram('vmaas_webapp_processing_seconds', 'Time spent processing request', ['method', 'path'])
+REQUEST_TIME = Histogram('vmaas_handler_processing_seconds', 'Time spent processing request', ['method', 'path'])
 
 # ...and then we'll build Counter for all-the-things into the BaseHandler
-REQUEST_COUNTS = Counter('vmaas_webapp_handler_invocations', 'Number of calls per handler',
+REQUEST_COUNTS = Counter('vmaas_handler_invocations', 'Number of calls per handler',
                          ['method', 'path', 'status'])
