+++ vmaas/webapp/app.py
@@ -3,16 +3,13 @@
 Main web API module
 """
 import asyncio
-import json
 import os
 import signal
 import re
-import time
 
 import connexion
 from prometheus_client import generate_latest
 import requests
-from starlette.datastructures import MutableHeaders
 from starlette.middleware.cors import CORSMiddleware
 from starlette.middleware.gzip import GZipMiddleware
 
@@ -24,8 +21,6 @@ from vmaas.webapp.packages import PackagesAPI
 from vmaas.webapp.pkglist import PkgListAPI
 from vmaas.webapp.patches import PatchesAPI
 from vmaas.webapp.pkgtree import PkgtreeAPI
-from vmaas.webapp.probes import REQUEST_COUNTS
-from vmaas.webapp.probes import REQUEST_TIME
 from vmaas.webapp.repos import RepoAPI
 from vmaas.webapp.rpm_pkg_names import RPMPkgNamesAPI
 from vmaas.webapp.srpm_pkg_names import SRPMPkgNamesAPI
@@ -36,6 +31,7 @@ from vmaas.common.config import Config
 from vmaas.common.constants import VMAAS_VERSION
 from vmaas.common.logging_utils import get_logger
 from vmaas.common.logging_utils import init_logging
+from vmaas.common.middlewares import ErrorHandlerMiddleware, TimingLoggingMiddleware
 from vmaas.common.strtobool import strtobool
 
 # pylint: disable=too-many-lines
@@ -504,81 +500,11 @@ def create_app(specs):
                     pass_context_arg_name='request',
                     arguments={'vmaas_version': VMAAS_VERSION})
 
-    class ErrorHandlerMiddleware:
-        """Middleware to wrap error message in JSON struct."""
-        def __init__(self, app):
-            self.app = app
-
-        def _format_error(self, detail, status):
-            res = {}
-            res["type"] = "about:blank"
-            res["detail"] = detail
-            res["status"] = status
-            return res
-
-        async def __call__(self, scope, receive, send):
-            if scope["type"] != "http":
-                await self.app(scope, receive, send)
-                return
-
-            start_message = None
-
-            async def send_with_formatted_error(message):
-                nonlocal start_message
-                if message["type"] == "http.response.start":
-                    start_message = message
-                elif message["type"] == "http.response.body":
-                    if start_message["status"] >= 400:
-                        headers = MutableHeaders(raw=start_message["headers"])
-                        error = message["body"].decode("utf-8").strip().replace('"', '')
-                        better_error = json.dumps(self._format_error(error, start_message["status"])).encode("utf-8")
-                        headers["Content-Length"] = str(len(better_error))
-                        message["body"] = better_error
-                    await send(start_message)
-                    await send(message)
-
-            await self.app(scope, receive, send_with_formatted_error)
-
-    class TimingMiddleware:
-        """Middleware to measure duration of requests to each endpoint and exporting it as Prometheus metric."""
-        def __init__(self, app):
-            self.app = app
-
-        async def __call__(self, scope, receive, send):
-            if scope["type"] != "http":
-                await self.app(scope, receive, send)
-                return
-
-            start_message = None
-            start_time = time.time()
-
-            async def send_finished(message):
-                nonlocal start_message
-                nonlocal start_time
-                if message["type"] == "http.response.start":
-                    start_message = message
-                elif message["type"] == "http.response.body":
-                    duration = time.time() - start_time
-                    method = scope["method"]
-                    path_parts = scope["path"].split('/')
-                    # (0) /(1) /(2)     /(3) /(4)
-                    #     /api (/vmaas) /v1  /updates
-                    if "vmaas" in path_parts:
-                        strip_limit = 5
-                    else:
-                        strip_limit = 4
-                    const_path = '/'.join(path_parts[:strip_limit])
-                    REQUEST_TIME.labels(method, const_path).observe(duration)
-                    REQUEST_COUNTS.labels(method, const_path, start_message["status"]).inc()
-                await send(message)
-
-            await self.app(scope, receive, send_finished)
-
     app.add_middleware(CORSMiddleware, allow_origins=["*"], allow_headers=["Content-Type"], allow_methods=["GET", "OPTIONS", "POST"])
     if GZIP_RESPONSE_ENABLE:
         app.add_middleware(GZipMiddleware, minimum_size=1, compresslevel=GZIP_COMPRESS_LEVEL)
     app.add_middleware(ErrorHandlerMiddleware)
-    app.add_middleware(TimingMiddleware)
+    app.add_middleware(TimingLoggingMiddleware, position=connexion.middleware.MiddlewarePosition.BEFORE_EXCEPTION)
 
     BaseHandler.db_cache = Cache()
     load_cache_to_apis()
